added fixed_point structure
The fixed_point class implements fixed-precision arithmetic. The precision is configured as a template parameter together with the number's storage type (int, short, etc.). The class provides handy conversions from/to other types, making it almost entirely transparent to the programmer. fixed_point may also be used as the data type of a vector.