added mvp, some transforms and fixes
Added orientation, perspective, and lookat transformations that perform the tasks of the rotational component of gluLookAt, gluPerspective and gluLookAt. Also added the mvp class which maintains a stack of model, view, and projection matrices. This allows very straightforward creation of the matrices required for OpenGL shaders. As an additional feature, the mvp class gives access to the intermediate model_view and model_view_projection, and the derived normal matrix. These are required for certain types of shaders, so it's handy to have them nearby.