added comparison ops to vector
The vector class now features <,>,<=,>=,==,!= operators. Arguably anything but the == and != operators does not make sense, as there is no proper definition of when a vector is to be considered less than or greater than another vector. Our implementation has sorting in a set or map in mind, and as such orders vectors by x,y,z,w components, in that order.